Add ActiveRate updater
