+++ src/cow-react/modules/limitOrders/updaters/InitialPriceUpdater/index.tsx
@@ -3,6 +3,7 @@ import { useUpdateAtom } from 'jotai/utils'
 import { updateLimitRateAtom } from '@cow/modules/limitOrders/state/limitRateAtom'
 import { useGetInitialPrice } from '@cow/modules/limitOrders/hooks/useGetInitialPrice'
 
+// Fetch and update initial price for the selected token pair
 export function InitialPriceUpdater() {
   const updateLimitRateState = useUpdateAtom(updateLimitRateAtom)
   const { price, isLoading } = useGetInitialPrice()
