+++ src/cow-react/pages/LimitOrders/index.tsx
@@ -6,12 +6,14 @@ import { Orders } from '@cow/modules/limitOrders/containers/Orders'
 import { InfoPopup } from '@cow/modules/limitOrders/pure/InfoPopup'
 import { QuoteUpdater } from '@cow/modules/limitOrders/updaters/QuoteUpdater'
 import { InitialPriceUpdater } from '@cow/modules/limitOrders/updaters/InitialPriceUpdater'
+import { ActiveRateUpdater } from '@cow/modules/limitOrders/updaters/ActiveRateUpdater'
 
 export default function LimitOrderPage() {
   return (
     <>
       <QuoteUpdater />
       <InitialPriceUpdater />
+      <ActiveRateUpdater />
       <styledEl.PageWrapper>
         <styledEl.Column>
           <LimitOrdersWidget />
